PRO-4611: Fixed after e2e framework restructure
