--- src/test/end-to-end/pages/eventSummary/grantOfProbate/addCommentSummaryConfig.json
@@ -1,5 +1,0 @@
-{
-  "waitForText" : "Add Comment",
-  "summary": "comment_event_summary_purpose_text",
-  "comment": "comment_event_description_text"
-}
