--- src/test/end-to-end/pages/eventSummary/grantOfProbate/caseMatchesCommentSummaryConfig.json
@@ -1,5 +1,0 @@
-{
-  "waitForText" : "Match application",
-  "summary": "match_application_event_summary_text",
-  "comment": "match_application_event_description_text"
-}
