--- src/test/end-to-end/pages/eventSummary/grantOfProbate/documentUploadSummaryConfig.json
@@ -1,5 +1,0 @@
-{
-  "waitForText" : "Upload Documents",
-  "summary": "upload_document_event_summary_purpose_text",
-  "comment": "upload_document_event_description_text"
-}
