--- src/test/end-to-end/pages/eventSummary/grantOfProbate/handleEvidenceSummaryConfig.json
@@ -1,6 +1,0 @@
-{
-    "waitForText" : "Handle supplementary evidence",
-    "summary": "handle_evidence_event_summary_purpose_text",
-    "comment": "handle_evidence_document_event_description_text"
-  }
-
