--- src/test/end-to-end/pages/eventSummary/grantOfProbate/markAsReadySummaryConfig.json
@@ -1,6 +1,0 @@
-{
-  "waitForText" : "Mark as ready for examination",
-  "summary": "mark_as_ready_event_summary_purpose_text",
-  "comment": "mark_as_ready_event_description_text"
-}
-
