--- src/test/end-to-end/pages/eventSummary/grantOfProbate/printCaseSummaryConfig.json
@@ -1,6 +1,0 @@
-{
-  "waitForText" : "Print the case",
-  "summary": "print_case_event_summary_purpose_text",
-  "comment": "print_case_document_event_description_text"
-}
-
