--- src/test/end-to-end/pages/eventSummary/grantOfProbate/withdrawalSummaryConfig.json
@@ -1,5 +1,0 @@
-{
-  "waitForText" : "Withdraw application",
-  "summary": "withdrawal_event_summary_purpose_text",
-  "comment": "withdrawal_event_description_text"
-}
