--- src/test/end-to-end/pages/eventSummary/willLodgement/addCommentSummaryConfig.json
@@ -1,5 +1,0 @@
-{
-  "waitForText" : "Add comment",
-  "summary": "comment_event_summary_purpose_text",
-  "comment": "comment_event_description_text"
-}
