--- src/test/end-to-end/pages/eventSummary/willLodgement/documentUploadSummaryConfig.json
@@ -1,5 +1,0 @@
-{
-  "waitForText" : "Upload document",
-  "summary": "upload_document_event_summary_text",
-  "comment": "upload_document_event_description_text"
-}
