+++ src/test/end-to-end/paths/caveatPath3.js
@@ -10,7 +10,7 @@ const createCaveatConfig = require('src/test/end-to-end/pages/createCaveat/creat
 const emailCaveatorConfig = require('src/test/end-to-end/pages/emailNotifications/caveat/emailCaveatorConfig');
 const reopenCaveatConfig = require('src/test/end-to-end/pages/reopenningCases/caveat/reopenCaveatConfig');
 const caseMatchesConfig = require('src/test/end-to-end/pages/caseMatches/caveat/caseMatchesConfig');
-const documentUploadConfig = require('src/test/end-to-end/pages/documentUpload/documentUploadConfig');
+const documentUploadConfig = require('src/test/end-to-end/pages/documentUpload/caveat/documentUploadConfig');
 
 const historyTabConfig = require('src/test/end-to-end/pages/caseDetails/caveat/historyTabConfig');
 
@@ -69,7 +69,7 @@ Scenario('Caveat Workflow - E2E Test 03 - Caveat for a Personal Applicant - Rais
 
     nextStepName = 'Upload document';
     I.chooseNextStep(nextStepName);
-    I.uploadDocument(caseRef);
+    I.uploadDocument(caseRef, documentUploadConfig);
     I.enterEventSummary(caseRef, nextStepName);
     // Note that End State does not change when uploading a document.
     I.seeCaseDetails(caseRef, historyTabConfig, eventSummaryConfig, nextStepName, endState);
