+++ src/test/end-to-end/paths/grantOfProbatePath.js
@@ -4,17 +4,9 @@ const testConfig = require('src/test/config');
 const createCaseConfig = require('src/test/end-to-end/pages/createCase/createCaseConfig');
 
 // const caseMatchesConfig = require('src/test/end-to-end/pages/caseMatches/grantOfProbate/caseMatchesConfig');
-const checkYourAnswersConfig = require('src/test/end-to-end/pages/checkYourAnswers/checkYourAnswersConfig');
 const createGrantOfProbateConfig = require('src/test/end-to-end/pages/createGrantOfProbate/createGrantOfProbateConfig');
 const documentUploadConfig = require('src/test/end-to-end/pages/documentUpload/grantOfProbate/documentUploadConfig');
-
-const addCommentSummaryConfig = require('src/test/end-to-end/pages/eventSummary/grantOfProbate/addCommentSummaryConfig');
-// const caseMatchesCommentSummaryConfig = require('src/test/end-to-end/pages/eventSummary/grantOfProbate/caseMatchesCommentSummaryConfig');
-const documentUploadSummaryConfig = require('src/test/end-to-end/pages/eventSummary/grantOfProbate/documentUploadSummaryConfig');
-const markAsReadySummaryConfig = require('src/test/end-to-end/pages/eventSummary/grantOfProbate/markAsReadySummaryConfig');
-const handleEvidenceSummaryConfig = require('src/test/end-to-end/pages/eventSummary/grantOfProbate/handleEvidenceSummaryConfig');
-const printCaseSummaryConfig = require('src/test/end-to-end/pages/eventSummary/grantOfProbate/printCaseSummaryConfig');
-const withdrawalSummaryConfig = require('src/test/end-to-end/pages/eventSummary/grantOfProbate/withdrawalSummaryConfig');
+const eventSummaryConfig = require('src/test/end-to-end/pages/eventSummary/eventSummaryConfig');
 
 const applicantDetailsTabConfig = require('src/test/end-to-end/pages/caseDetails/grantOfProbate/applicantDetailsTabConfig');
 const caseDetailsTabConfig = require('src/test/end-to-end/pages/caseDetails/grantOfProbate/caseDetailsTabConfig');
@@ -47,7 +39,7 @@ Scenario('Grant of Probate Workflow - E2E test 01 - Grant of Representation for 
     I.enterGrantOfProbatePage7('create');
     I.enterGrantOfProbatePage8('create');
     I.enterGrantOfProbatePage9('create');
-    I.checkMyAnswers();
+    I.checkMyAnswers(nextStepName);
     let endState = 'Case created';
 
     const url = await I.grabCurrentUrl();
@@ -55,7 +47,7 @@ Scenario('Grant of Probate Workflow - E2E test 01 - Grant of Representation for 
         .match(/.{4}/g)
         .join('-');
 
-    I.seeCaseDetails(caseRef, historyTabConfig, checkYourAnswersConfig, nextStepName, endState);
+    I.seeCaseDetails(caseRef, historyTabConfig, eventSummaryConfig, nextStepName, endState);
     I.seeCaseDetails(caseRef, deceasedTabConfig, createGrantOfProbateConfig);
     I.seeCaseDetails(caseRef, caseDetailsTabConfig, createGrantOfProbateConfig);
     I.seeCaseDetails(caseRef, applicantDetailsTabConfig, createGrantOfProbateConfig);
@@ -64,41 +56,41 @@ Scenario('Grant of Probate Workflow - E2E test 01 - Grant of Representation for 
     nextStepName = 'Handle supplementary evidence';
     I.chooseNextStep(nextStepName);
     I.handleEvidence(caseRef);
-    I.enterEventSummary(caseRef, handleEvidenceSummaryConfig);
+    I.enterEventSummary(caseRef, nextStepName);
     endState = 'Case created';
-    I.seeCaseDetails(caseRef, historyTabConfig, handleEvidenceSummaryConfig, nextStepName, endState);
+    I.seeCaseDetails(caseRef, historyTabConfig, eventSummaryConfig, nextStepName, endState);
 
     nextStepName = 'Amend case details';
     I.chooseNextStep(nextStepName);
     I.enterGrantOfProbatePage1('update');
-    I.checkMyAnswers();
+    I.checkMyAnswers(nextStepName);
     I.chooseNextStep(nextStepName);
     I.enterGrantOfProbatePage2('update');
-    I.checkMyAnswers();
+    I.checkMyAnswers(nextStepName);
     I.chooseNextStep(nextStepName);
     I.enterGrantOfProbatePage3('update');
-    I.checkMyAnswers();
+    I.checkMyAnswers(nextStepName);
     I.chooseNextStep(nextStepName);
     I.enterGrantOfProbatePage4('update');
-    I.checkMyAnswers();
+    I.checkMyAnswers(nextStepName);
     I.chooseNextStep(nextStepName);
     I.enterGrantOfProbatePage5('update');
-    I.checkMyAnswers();
+    I.checkMyAnswers(nextStepName);
     I.chooseNextStep(nextStepName);
     I.enterGrantOfProbatePage6('update');
-    I.checkMyAnswers();
+    I.checkMyAnswers(nextStepName);
     I.chooseNextStep(nextStepName);
     I.enterGrantOfProbatePage7('update');
-    I.checkMyAnswers();
+    I.checkMyAnswers(nextStepName);
     I.chooseNextStep(nextStepName);
     I.enterGrantOfProbatePage8('update');
-    I.checkMyAnswers();
+    I.checkMyAnswers(nextStepName);
 
     // I.chooseNextStep(nextStepName);
     // I.enterGrantOfProbatePage9('update');
-    // I.checkMyAnswers();
+    // I.checkMyAnswers(nextStepName);
 
-    I.seeCaseDetails(caseRef, historyTabConfig, checkYourAnswersConfig, nextStepName, endState);
+    I.seeCaseDetails(caseRef, historyTabConfig, eventSummaryConfig, nextStepName, endState);
     I.seeCaseDetails(caseRef, deceasedUpdateTabConfig, createGrantOfProbateConfig);
     I.seeCaseDetails(caseRef, caseDetailsUpdateTabConfig, createGrantOfProbateConfig);
     I.seeCaseDetails(caseRef, applicantDetailsUpdateTabConfig, createGrantOfProbateConfig);
@@ -106,40 +98,40 @@ Scenario('Grant of Probate Workflow - E2E test 01 - Grant of Representation for 
     nextStepName = 'Print the case';
     I.chooseNextStep(nextStepName);
     I.printCase(caseRef);
-    I.enterEventSummary(caseRef, printCaseSummaryConfig);
+    I.enterEventSummary(caseRef, nextStepName);
     endState = 'Awaiting documentation';
-    I.seeCaseDetails(caseRef, historyTabConfig, printCaseSummaryConfig, nextStepName, endState);
+    I.seeCaseDetails(caseRef, historyTabConfig, eventSummaryConfig, nextStepName, endState);
 
     nextStepName = 'Add Comment';
     I.chooseNextStep(nextStepName);
-    I.enterComment(caseRef, addCommentSummaryConfig);
-    I.seeCaseDetails(caseRef, historyTabConfig, addCommentSummaryConfig, nextStepName, endState);
+    I.enterComment(caseRef, nextStepName);
+    I.seeCaseDetails(caseRef, historyTabConfig, eventSummaryConfig, nextStepName, endState);
 
     nextStepName = 'Upload Documents';
     I.chooseNextStep(nextStepName);
     I.uploadDocument(caseRef, documentUploadConfig);
-    I.enterEventSummary(caseRef, documentUploadSummaryConfig);
-    I.seeCaseDetails(caseRef, historyTabConfig, documentUploadSummaryConfig, nextStepName, endState);
+    I.enterEventSummary(caseRef, nextStepName);
+    I.seeCaseDetails(caseRef, historyTabConfig, eventSummaryConfig, nextStepName, endState);
     I.seeCaseDetails(caseRef, documentUploadTabConfig, documentUploadConfig);
 
     nextStepName = 'Mark as ready for examination';
     I.chooseNextStep(nextStepName);
     I.markAsReadyForExamination(caseRef);
-    I.enterEventSummary(caseRef, markAsReadySummaryConfig);
+    I.enterEventSummary(caseRef, nextStepName);
     endState = 'Ready for examination';
-    I.seeCaseDetails(caseRef, historyTabConfig, markAsReadySummaryConfig, nextStepName, endState);
+    I.seeCaseDetails(caseRef, historyTabConfig, eventSummaryConfig, nextStepName, endState);
 
     // nextStepName = 'Find matches (Examining)';
     // I.chooseNextStep(nextStepName);
     // I.selectCaseMatchesForGrantOfProbate(caseRef, caseMatchesConfig);
-    // I.enterCaseMatchesComment(caseRef, caseMatchesCommentSummaryConfig);
-    // I.seeCaseDetails(caseRef, historyTabConfig, caseMatchesCommentSummaryConfig, nextStepName, endState);
+    // I.enterEventSummary(caseRef, nextStepName);
+    // I.seeCaseDetails(caseRef, historyTabConfig, ceventSummaryConfig, nextStepName, endState);
     // I.seeCaseDetails(caseRef, caseMatchesTabConfig, caseMatchesConfig);
 
     nextStepName = 'Withdraw application';
     I.chooseNextStep(nextStepName);
-    I.enterWithdrawalSummary(caseRef, withdrawalSummaryConfig);
+    I.enterEventSummary(caseRef, nextStepName);
     endState = 'Case closed';
-    I.seeCaseDetails(caseRef, historyTabConfig, withdrawalSummaryConfig, nextStepName, endState);
+    I.seeCaseDetails(caseRef, historyTabConfig, eventSummaryConfig, nextStepName, endState);
 
 }).retry(testConfig.TestRetryScenarios);
